feat: Add timer feature to Output class
This commit adds a timer feature to the Output class in the output.py file. The timer function runs every 3 seconds and can be used to create timers. This feature allows the user to set a timer for duration and execute a command after the timer ends. This commit also includes some minor changes to improve code readability and maintainability.